Add features
Create a singleton cache that can be updated from the fast api layer if a user would like to remove a token explicitly from the cache.
Enhance documentation
Allow the roles_key to be set at the middleware and the secure_route wrapper.
update to prefer single quotes where possible within the code as it is today.