Modifying implementation to be a little more flexible
modifying implementation to support a logger protocol so the library doesn't care what the logger is as long as it matches the logger protocol at runtime.
modify the request context so that it will propagate the request context to the wrapper instead of requiring that the route defines the request: Request object in its signature.
Updating some code comments so they are clearer regarding what they are doing. Updating doc comments as well to be clearer for maintainablity.
updating pyproject.toml to exclude tests in the package.
updated debug logs and added more to the caching layer.