update tests so they run without errors
There is a race condition in the execution of tests if the token_cache and token_cache_singleton occur at the same time. To avoid this I put the token_cache_singleton tests at the beginning of the test runs and the token_cache tests at the end of the test runs. This is only an issue when running from the CLI, not from the built in test runners.